[refactor] change directory structure
