+++ src/main/java/app/Application.java
@@ -1,13 +1,13 @@
-package com.ranly.book;
+package app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BookApplication {
+public class Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BookApplication.class, args);
+		SpringApplication.run(Application.class, args);
 	}
 
 }
