+++ src/main/java/app/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com.ranly.book;
+package app;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
@@ -7,7 +7,7 @@ public class ServletInitializer extends SpringBootServletInitializer {
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(BookApplication.class);
+		return application.sources(Application.class);
 	}
 
 }
