+++ src/main/java/app/controller/BookController.java
@@ -1,4 +1,4 @@
-package com.ranly.book.controller;
+package app.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -6,8 +6,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class BookController {
-    @RequestMapping(method = RequestMethod.GET, value = "/books")
-    public String getBooks() {
-        return "Books";
-    }
+
+	@RequestMapping(method = RequestMethod.GET, value = "/book")
+	public String getBook() {
+		return "Book";
+	}
+
 }
