+++ src/main/java/app/controller/ServiceRootController.java
@@ -1,4 +1,4 @@
-package com.ranly.book.controller;
+package app.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -9,9 +9,16 @@ import com.google.gson.JsonObject;
 @RestController
 public class ServiceRootController {
 	@RequestMapping(method = RequestMethod.GET, value = "/")
-	public String getEntry() {
+	public String getServiceRoot() {
 		JsonObject entry = new JsonObject();
-		entry.addProperty("Path", "/books");
+		entry.addProperty("Path", "/api/v1");
+		return entry.toString();
+	}
+
+	@RequestMapping(method = RequestMethod.GET, value = "/api/v1")
+	public String getApiEntry() {
+		JsonObject entry = new JsonObject();
+		entry.addProperty("Path", "/api/v1/book");
 		return entry.toString();
 	}
 }
