+++ src/test/java/app/ApplicationTests.java
@@ -1,10 +1,10 @@
-package com.ranly.book;
+package app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BookApplicationTests {
+class ApplicationTests {
 
 	@Test
 	void contextLoads() {
